all: update to Gio 0.6 and Go 1.22
That change uses the new gtx.Execute. This change makes
compatible with Gio 0.6. The plugins.Install was removed
and replaced by gioplugins.Event(), README was updated
to reflect such change.

Also, the wasm-import and InkWASM to be compatible with
Go 1.20+ versions. The Explorer package was updated to
be compatible with standard version of Go Compiler.

Fix #57
Fix #65

Signed-off-by: inkeliz <inkeliz@inkeliz.com>