all: [API CHANGE] update to Gio 0.6 and Go 1.22 (#69)
* all: update to Gio 0.6 and Go 1.22

That change uses the new gtx.Execute. This change makes
compatible with Gio 0.6. The plugins.Install was removed
and replaced by gioplugins.Hijack(), README was updated
to reflect such change. 

Also, you need to use special functions to access events,
such as `gioplugins.Event(gtx, {filter})`. It's no longer
compatible with gtx.Event/gtx.Execute.

Also, the wasm-import and InkWASM to be compatible with
Go 1.20+ versions. The Explorer package was updated to
be compatible with standard version of Go Compiler.

Fix #57
Fix #65

Signed-off-by: inkeliz <inkeliz@inkeliz.com>

* all: add Makefile for demo and generator

Signed-off-by: inkeliz <inkeliz@inkeliz.com>

* plugin: improve swap

Signed-off-by: inkeliz <inkeliz@inkeliz.com>

* all: update generated files

Signed-off-by: inkeliz <inkeliz@inkeliz.com>

* all: make compatible with std Gio 0.6

Signed-off-by: inkeliz <inkeliz@inkeliz.com>

---------

Signed-off-by: inkeliz <inkeliz@inkeliz.com>